+++ OpenAI.SDK/ObjectModels/ResponseModels/FineTuningJobResponseModels/FineTuningJobResponse.cs
@@ -67,7 +67,7 @@ public record FineTuningJobResponse : BaseResponse, IOpenAIModels.IId, IOpenAIMo
     ///     The Unix timestamp (in seconds) for when the fine-tuning job was created.
     /// </summary>
     [JsonPropertyName("created_at")]
-    public int CreatedAt { get; set; }
+    public long CreatedAt { get; set; }
 
     /// <summary>
     ///     The object identifier, which can be referenced in the API endpoints.
